tidy info struct again, add platform detection
There is a new InstallPlatform() method for info structs that returns
the Instructions struct (also a new refactor) for the respective
platform, detected via runtime.GOOS.

Essentially, instead of looking at Info.InstallConf.Linux, we use
Info.InstallConf.InstallPlatform() to select the correct sub-struct
holding installation instructions.

If this ends up being a good idea, I assume we should mirror this
implementation for something like UninstallPlatform(). But uninstalling
is still to be implemented.